Fix: Address erratic behavior in Cypress tests
Currently experiencing inconsistent behavior in the Cypress tests. They seem to take longer than usual to load and exhibit unpredictable results. This commit serves as a record of these issues for future investigation, as I'm unsure if it's related to the Cypress service or the test setup. Further review and adjustments will be needed down the line.